refactor(plug): use dynamic import for subcommand
